+++ src/os/pl-fmt.c
@@ -452,7 +452,7 @@ do_format(IOSTREAM *fd, PL_chars_t *fmt, int argc, term_t argv, Module m)
 	    int i;
 
 	    PL_predicate_info(proc, NULL, &arity, NULL);
-	    av = PL_new_term_refs(arity);
+	    av = PL_new_term_refs((int)arity);
 
 	    if ( arg == DEFAULT )
 	      PL_put_atom(av+0, ATOM_default);
@@ -1167,7 +1167,7 @@ utf8_dp(PL_locale *l, char *s, int *len)
 { if ( l->decimal_point )
   { if ( !ths_to_utf8(s, l->decimal_point, 20) )
       return FALSE;
-    *len = strlen(s);
+    *len = (int)strlen(s);
   } else
   { *s++ = '.';
     *s = EOS;
@@ -1248,7 +1248,7 @@ groupDigits(PL_locale *locale, Buffer b)
 
       if ( !ths_to_utf8(ths, locale->thousands_sep, sizeof(ths)) )
 	return FALSE;
-      thslen = strlen(ths);
+      thslen = (int)strlen(ths);
 
       if ( !growBuffer(b, thslen*groups) )
 	return PL_no_memory();
